requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pandas>=2.0.0
pydantic>=2.0.0
tiktoken>=0.7.0
